Fixes for ZFS subpools
