+++ emhttp/plugins/dynamix/include/Helpers.php
@@ -15,10 +15,6 @@ $docroot ??= ($_SERVER['DOCUMENT_ROOT'] ?: '/usr/local/emhttp');
 require_once "$docroot/webGui/include/Wrappers.php";
 require_once "$docroot/webGui/include/Secure.php";
 
-// ZFS subpool name separator and replacement
-$tilde = '~';
-$proxy = '__';
-
 // Helper functions
 function my_scale($value, &$unit, $decimals=NULL, $scale=NULL, $kilo=1000) {
   global $display,$language;
@@ -66,13 +62,6 @@ function my_disks($disk) {
 }
 function my_hyperlink($text, $link) {
   return str_replace(['[',']'],["<a href=\"$link\">","</a>"],$text);
-}
-function no_tilde($name) {
-  global $tilde,$proxy;
-  return str_replace($tilde,$proxy,$name);
-}
-function prefix($key) {
-  return preg_replace('/\d+$/','',$key);
 }
 function parity_only($disk) {
   return _var($disk,'type')=='Parity';
